Ghost Hitbox modified
in this update i resized the hitbox on the ghost to make it a bit smarter in regards to obstacle avoidance.

I will likely be taking a break from this scene to begin working on another one so i can comeback to this when i get a better understanding of how the animation process works.